move helpers to VectorManager
