+++ libs/server/Resp/Vector/VectorManager.Migration.cs
@@ -3,6 +3,7 @@
 
 using System;
 using System.Buffers.Binary;
+using System.Collections.Frozen;
 using System.Collections.Generic;
 using System.Diagnostics;
 using System.Runtime.InteropServices;
@@ -333,7 +334,7 @@ namespace Garnet.server
         /// 
         /// Meant for use during migration.
         /// </summary>
-        public unsafe HashSet<ulong> GetNamespacesForKeys(StoreWrapper storeWrapper, IEnumerable<PinnedSpanByte> keys, Dictionary<byte[], byte[]> vectorSetKeys)
+        public HashSet<ulong> GetNamespacesForKeys(StoreWrapper storeWrapper, IEnumerable<PinnedSpanByte> keys, Dictionary<byte[], byte[]> vectorSetKeys)
         {
             // TODO: Ideally we wouldn't make a new session for this, but it's fine for now
             using var storageSession = new StorageSession(storeWrapper, new(), new(), null, null, storeWrapper.DefaultDatabase.Id, null, this, logger);
@@ -368,6 +369,155 @@ namespace Garnet.server
             }
 
             return namespaces;
+        }
+
+
+        /// <summary>
+        /// Update the namespaces stored in <paramref name="readOutput"/> according to <see cref="FrozenDictionary"/>.
+        /// 
+        /// <paramref name="readInput"/> should have been used to populate <paramref name="readOutput"/> with a Tsavorite Read prior to this call.
+        /// </summary>
+        public static void UpdateMigratedElementNamespaces(FrozenDictionary<ulong, ulong> oldToNewNamespaces, ref VectorInput readInput, ref VectorOutput readOutput)
+        {
+            Debug.Assert(readInput.IsMigrationRead, "Unexpected input");
+
+            DeserializeMigratedElementKey(readOutput.SpanByteAndMemory.Span, out var namespaceBytes, out _, out _);
+
+            ulong oldNs = BinaryPrimitives.ReadUInt32LittleEndian(namespaceBytes);
+
+            if (!oldToNewNamespaces.TryGetValue(oldNs, out var newNs))
+            {
+                return;
+            }
+
+            Debug.Assert(newNs <= uint.MaxValue, "Shouldn't have reserved such a large context");
+
+            BinaryPrimitives.WriteUInt32LittleEndian(namespaceBytes, (uint)newNs);
+        }
+
+        /// <summary>
+        /// Calculate needed storage for migrating a Vector Set element key.
+        /// </summary>
+        public static int GetMigratedElementKeySerializationSize(ReadOnlySpan<byte> keyBytes, ReadOnlySpan<byte> alignedValue)
+        {
+            var neededSpace =
+                sizeof(int) + sizeof(uint) + // Namespace is ALWAYS expanded to 4-bytes so we have space to re-write it
+                sizeof(int) + keyBytes.Length +
+                sizeof(int) + alignedValue.Length;
+
+            return neededSpace;
+        }
+
+        /// <summary>
+        /// Serialize a record for migrating a Vector Set element key.
+        /// </summary>
+        public static void SerializeMigratedElementKey(Span<byte> dataBytes, ReadOnlySpan<byte> namespaceBytes, ReadOnlySpan<byte> keyBytes, ReadOnlySpan<byte> alignedValue)
+        {
+            ulong context;
+            if (namespaceBytes.Length == 1)
+            {
+                context = namespaceBytes[0];
+            }
+            else
+            {
+                Debug.Assert(namespaceBytes.Length == 4, "Unexpected namespace size");
+                context = BinaryPrimitives.ReadUInt32LittleEndian(namespaceBytes);
+            }
+
+            var writeTo = dataBytes;
+
+            // Namespace length, Namespace
+            BinaryPrimitives.WriteInt32LittleEndian(writeTo, 4);
+            writeTo = writeTo[sizeof(int)..];
+            BinaryPrimitives.WriteUInt32LittleEndian(writeTo, (uint)context);
+            writeTo = writeTo[sizeof(uint)..];
+
+            // Key length, Key
+            BinaryPrimitives.WriteInt32LittleEndian(writeTo, keyBytes.Length);
+            writeTo = writeTo[sizeof(int)..];
+            keyBytes.CopyTo(writeTo);
+            writeTo = writeTo[keyBytes.Length..];
+
+            // Value length, Value
+            BinaryPrimitives.WriteInt32LittleEndian(writeTo, alignedValue.Length);
+            writeTo = writeTo[sizeof(int)..];
+            alignedValue.CopyTo(writeTo);
+        }
+
+        /// <summary>
+        /// Reverse <see cref="SerializeMigratedElementKey"/>.
+        /// </summary>
+        public static void DeserializeMigratedElementKey(Span<byte> dataBytes, out Span<byte> namespaceBytes, out Span<byte> keyBytes, out Span<byte> value)
+        {
+            var readFrom = dataBytes;
+
+            // Namespace length, Namespace
+            var nsLength = BinaryPrimitives.ReadInt32LittleEndian(readFrom);
+            Debug.Assert(nsLength == sizeof(uint), "Namespace should always be 4-bytes when deserializing");
+            readFrom = readFrom[sizeof(uint)..];
+            namespaceBytes = readFrom[..nsLength];
+            readFrom = readFrom[nsLength..];
+
+            // Key length, Key
+            var keyLength = BinaryPrimitives.ReadInt32LittleEndian(readFrom);
+            readFrom = readFrom[sizeof(int)..];
+            keyBytes = readFrom[..keyLength];
+            readFrom = readFrom[keyLength..];
+
+            // Value length, Value
+            var valueLength = BinaryPrimitives.ReadInt32LittleEndian(readFrom);
+            readFrom = readFrom[sizeof(int)..];
+            value = readFrom[..valueLength];
+        }
+
+        /// <summary>
+        /// Calculate needed storage for migrating a Vector Set index key.
+        /// </summary>
+        public static int GetMigratedIndexKeySerializationSize(ReadOnlySpan<byte> keyBytes, ReadOnlySpan<byte> valueBytes)
+        {
+            var neededSpace = sizeof(int) + keyBytes.Length + sizeof(int) + valueBytes.Length;
+
+            return neededSpace;
+        }
+
+        /// <summary>
+        /// Serialize a record for migrating a Vector Set index key.
+        /// </summary>
+        public static void SerializeMigratedIndexKey(Span<byte> dataBytes, ReadOnlySpan<byte> keyBytes, ReadOnlySpan<byte> valueBytes)
+        {
+            Debug.Assert(valueBytes.Length == VectorManager.IndexSize, "Should only ever serialize index");
+
+            var writeTo = dataBytes;
+
+            // Key length, Key
+            BinaryPrimitives.WriteInt32LittleEndian(writeTo, keyBytes.Length);
+            writeTo = writeTo[sizeof(int)..];
+            keyBytes.CopyTo(writeTo);
+            writeTo = writeTo[keyBytes.Length..];
+
+            // Value length, Value
+            BinaryPrimitives.WriteInt32LittleEndian(writeTo, valueBytes.Length);
+            writeTo = writeTo[sizeof(int)..];
+            valueBytes.CopyTo(writeTo);
+        }
+
+        /// <summary>
+        /// Reverse <see cref="SerializeMigratedIndexKey"/>.
+        /// </summary>
+        public static void DeserializeMigratedIndexKey(ReadOnlySpan<byte> dataBytes, out ReadOnlySpan<byte> keyBytes, out ReadOnlySpan<byte> valueBytes)
+        {
+            var readFrom = dataBytes;
+
+            // Key length, Key
+            var keyLength = BinaryPrimitives.ReadInt32LittleEndian(readFrom);
+            readFrom = readFrom[sizeof(int)..];
+            keyBytes = readFrom[..keyLength];
+            readFrom = readFrom[keyLength..];
+
+            // Value length, Value
+            var valueLength = BinaryPrimitives.ReadInt32LittleEndian(readFrom);
+            readFrom = readFrom[sizeof(int)..];
+            valueBytes = readFrom[..valueLength];
         }
     }
 }
